common: linux: allow override of ENABLE_COLOR
The way a thread color is chosen based on pthread_self() depends very
much on the pthread implementation: it assumes pthread_t is a PID. Some
pthread implementations (e.g. musl) make pthread_t a pointer, which
breaks the assumptions.

Simply allow ENABLE_COLOR to be specified on the command line.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>