ale_tests: test infra for al_entity
The test infrastructure creates 4 veth interfaces and starts al_entity
with the simulated mode on each interface. Then it starts a test
executable that can interact with those interfaces.

All of this requires superuser access, so a wrapper script sets up
the virtual interfaces as root and tears them down again.

For now, only a nop test is included.

The tests are launched from .travis.yml. Since we want to run the
ale_tests even if the unit_tests fail, we need to use a different
syntax to sequence the steps in the script. Even if the al_entity
fails to build, we still run the unit tests. And even if the unit tests
fail, we still run the ale_tests.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>