ale_tests: start_interfaces: clean exit on error
If the shell gets killed, we still want to kill the al_entity and bring
down the virtual interfaces. Register a trap function to do this.

dash apparently doesn't do this correctly so run the script in bash.

The test executable is now the last command in the script so its exit
code is also the exit code of the script - which is what we want of
course. Before, the script always exited successfully (at least if
it was possible to bring down the virtual interfaces). To avoid this
mistake in the future, do an explicit exit if the test executable fails.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>