al: add draft interface.h header
The goal is to use a new struct interface with callbacks to handle
the interface-specific differences. This will allow to have different
implementations of querying or controlling the interface. For example,
one interface might use cfg80211, while another interface might use
something proprietary.

This is just a draft header, there is no implementation yet and it's
not used anywhere.