AL: use packet socket instead of pcap
Use a socket opened with openPacketSocket() to receive packets instead
of relying on pcap.

The AL and multicast addresses must be added to the interface,
otherwise no packet will be received or sent over it. To make sure that
the addresses have been added before we start sending packets, a sleep
is added after starting the interface thread. This is of course a
horrible hack, but in the long term the idea is to move the opening of
the sockets and setting of addresses to the interface initialisation
itself.

The byte-per-byte multicast address definition is no longer needed, so
define MCAST_1905 and MCAST_LLDP as a string instead of an array. That
way it can be used without local variable.

Remove pcap from travis, README and linking in the Makefile.