common: packet_tools: add _[EI]xBL functions
When extracting/inserting, we *must* always check that the buffer length
is not exceeded. Just like we keep a running packet_ppointer, we can
keep the remaining length and update it whenever something
extracted/inserted.

The new functions use the new coding style for types: bool, size_t and
uint*_t. Also the pointers are properly constified. Since the
underlying _[EI]xB functions are not yet constified, an explicit cast
of packet_ppointer is needed.