factory: 1905_cmdus: refactor the TLV count check
Currently the TLV count check is done with a bitmask. That works
because only 30 TLV types are used. However, EasyMesh adds more than 20
new TLV types, and their number starts at 128, so a bitmask is not so
convenient anymore.

Therefore, convert the bitmask into a list. We take the opportunity to
switch from separate arrays per check to a single array that specifies
the required counts for each TLV. For that, a new enum count_required
is introduced, and this is used in a new struct
cmdu_tlv_count_required. Conveniently, the 0 value of enum
count_required can be used as a sentinel marking the end of the list.

With this new type, it is also easy to add checks for the zero_or_one
and one_or_more constraints that up to now were not checked for
correctly.

Since we're anyway refactoring, also introduce struct cmdu_info that
will allow us to gather all CMDU constraints in one place. E.g. the
relaying information can be added to that structure as well - but that
is for later refactoring steps.