common: remove PLATFORM_MEMCMP
Use memcmp() (and include string.h) directly.

Note that there was a difference between PLATFORM_MEMCMP() and memcmp():
PLATFORM_MEMCMP() would return 1 instead of -1 if s1 > s2. However, all
users of PLATFORM_MEMCMP either compare to 0 or use the return value
directly as a boolean, so this doesn't make a difference.