common: rename PLATFORM_MALLOC to memalloc and move to utils
malloc() is a standard library function so we can use it directly.
However, we want to keep the additional functionality to exit on failure.
Therefore, rename it to memalloc().

It is not a platform-related function so move it to utils.h. Also it is
an excellent candidate for inlining.

While we're at it, send the error output to stderr instead of stdout.