common/hlist: new basic data structure
A TLV is essentially a hierarchical structure, with lists of
substructures. By abstracting this into the hlist structure, we can
avoid the need for type-specific free and compare functions. Also,
a linked list makes it easier to construct the TLV (or CMDU).

hlist is in common, but its unit test is in factory because common
currently doesn't have unit tests.

Also add contianer_of() and check_compatible_types() to common/utils.h