packet_tools: constify _ExB()
The packet argument of the _ExB() functions is not modified so it should
be const.

This requires propagating the const to a whole lot of other functions
and variables, because const uint8_t** and uint8_t** are incompatible.

The propagation is pushed through a little further than striclty
necessary, since we anyway want things to be const whenever possible.