factory: move print and compare functionality to struct tlv
hlist was not the right place for it.

We add an intermediate class tlv_struct to model the structures used
in TLVs. It is these structures that can be printed and compared.

The tests for printing and comparing are moved to 1905_tlv_forging.
It's convenient to put it there since we already have the infra to
create 1905 TLVs.