al: introduce new datamodel
The new datamodel is based on a hierarchical structure of devices,
interfaces, etc. Compared to the existing model, it has the following
purposes:
* Based on hlist, which simplifies dynamic updates of the model.
* Structures are directly accessible, reducing the need for copying
  information to return values.
* Subclassing for interface-specific information and callbacks.
* Remove dependency on 1905 TLVs; the dependency may be reverted later,
  i.e. 1905 TLV substructures may use structures defined in datamodel.h.

For now, al_datamodel is still used as well. Objects in al_datamodel
will be gradually converted to the new datamodel objects, and the
al_datamodel accessor functions will either use the new datamodel
objects, or be removed in favour of accessor functions that use the
new datamodel directly.

At this time, only the local AL MAC address is converted.