al/datamodel: add WSC and registrar model
Instead of getting the info from the single interface that acts as
registrar, allow WSC to be configured independently of the interfaces.
For now, the config is still taken from the interface but that can be
changed later.

The same structure will also be used to store the information received
from the registrar, but for now it's only used when the local device is
the registrar.

While we're at it, fix a few small issues: the arrays for manufacturer
etc. were 1 byte short; 11af interfaces would use an rf_band field of
0 (now they give an error).