al: datamodel: use neighbour devices from new datamodel
This is a stupid substitution without optimising the functions in
al_datamodel.c to make use of the better representation.

For now, the model is wrong because the same neighbor device occurs
several times if there are several links to it.

Since the hlist points to itself, realloc() doesn't work anymore.
Therefore, for the time being, we statically allocate the array of
local interfaces.