al: rework the datamodel
All the topology information is now recorded in the data model.

Devices and interfaces are tracked as dlists. Neighbors are tracked as
pointer arrays, since a single interface can be neighbor to multiple
other interfaces.

Interfaces can have no owner (i.e. non-1905 interfaces); however, this
is not used at the moment.

Most of the datamodel functions are moved to datamodel itself instead
of al_datamodel. The rest of the code is still calling the al_datamodel
functions, however.