platform_interfaces: use datamodel
Add createLocalInterfaces() function that creates the interfaces in the
datamodel. The datamodel is filled in in exactly the same way as it was
before, except that the wrt1900acx handling is removed entirely.
PLATFORM_GET_1905_INTERFACE_INFO() is rewritten to use the datamodel
where possible. Now, it will be possible to replace calls to this
function with direct accesses to the data model.

Power state is added to the data model.

Extensions are still handled like they were before.