Handle the Supported and Searched Service TLVs
In AP Autoconfiguration Search, we just ignored the Supported and
Searched service TLVs. Fix that: a Multi-AP response is only sent of the
Searched Service is Controller, and keep track of the Supported services
based on the Supported Service TLV. For the latter, we keep track in the
data model if the device is a controller and/or agent.

In AP Autoconfiguration Response, we didn't even handle the Supported
Service TLV. There also we use it to update the model.

In both of these, add the sending device to the data model if it was
not there yet.

While we're at it, refactor things a little and also constify the
arguments of the send functions.