al_recv: AP Autoconfiguration Response: use radios instead of interfaces
In Multi-AP, we need to send a WSC message when the radio is
unconfigured, i.e. when there is no interface on it. So instead of
iterating over interfaces, iterate over radios and check if they have a
configured bss.

While we're at it, use struct radio instead of
PLATFORM_GET_1905_INTERFACE_INFO().