al_send: send AP Radio Basic Capabilities
It has to be sent in the M1 AP Autoconfiguration WSC CMDU. It is only
sent if the registrar is a Multi-AP controller, as determined by the
supported services.

For the time being, supported channel classes are not included. These
should be constructed somehow from the available channel lists, but
that's quite complicated.

Note that no testing is included.