al_entity: start1905AL: iterate over datamodel instead of interface names
DMinsertInterface() is no longer needed, the interfaces are already
there.

The "Obtain the list of interfaces" loop is no longer needed; it is only
used to get the registrar interface now, so use that one directly.

Where possible, use struct interface members instead of going through
the old struct interfaceInfo.