al_wsc: rework handling of M1
Instead of storing M1 in a local variable, keep it in the radio struct.
This allows us to have several M1 active at the same time (though
that is a problem for non-multi-AP because M2 doesn't specify which
radio it applies to). It also makes it a lot clearer what is happening.

For the time being, the device info in M1 is just dummy data (before
this commit, it was actually invalid because we tried to receive it
from the radio). Ideally, this should be set in the radio structure.