al_entity: send multiple WSC M2 frames, one per SSID
The controller may be configured with multiple SSIDs. In that case,
each SSID that matches the radio's capabilities should be provided
in a WSC TLV containing M2 frame in the AP Autoconfiguration WSC CMDU.
Therefore, find all SSIDs that match the radio's rf_band and construct
an M2 for them. The M2s are collected in a new ptrarray type wscM2List.

Since the way that an M1 and an M2 WSC CMDU gets constructed are now
fairly different, so we separate into
send1905APAutoconfigurationWSCM1Packet() and
send1905APAutoconfigurationWSCM2Packet().