platform_uci: free the uciradiolist objects
In the reghandlers_out path of uci_register_handlers(), we free() the
strings in the uciradiolist structs, but not the uciradiolist structs
themselves. Fix this. We can then no longer iterate over the list in the
normal way because the just-freed object would still be used in the
for_each. So instead, pop the first element and remove it from the list.

Reported-by: Daniel Golle <daniel@makrotopia.org>